internal/generate/collector: refactor manifestCollection here and export as Manifests, but not apply* methods which will stay in CSV package
internal/generate/olm-catalog: use collector.Manifests
